Implemented a factory pattern for creating TP managers, and split it into two classes: one that's readonly blockchain methods, and one that has write capabilities. The factory method returns the appropriate instance depending on the client that has been provided. If the client has the necessary information to sign transactiosn on the blockchain, a writeable TP maanger will be returned -- otherwise a readonly one will be returned.
With TP managers, a readonly manager is used by saying `manager.read.someFunc`, which makes it clear that the method being called is a read method. For writeable managers, you say `manager.write.role.someFunc`, which not only specifies that the method is a write method, but also which TandaPay role it is associated, so it's easy to differentiate between secretary-only and all-member methods. This cuts down on confusion dramatically.

Finally, an options and .extend pattern has been implemented for writeable TP manager instances. It can be constructed with options that specify which role the client has in the SC instance, i.e. member or secretary, and then only the appropriate methods will be available. It can also be extended after construction, such as with .extend(MemberWriteActions) or .extend(SecretaryWriteActions) (... identifiers subject to change)