**finally** successfully got accounts to be able to pay their premiums.
NOTE: There is an issue where the tests sometimes randomly fail, this seems to be a problem with anvil and not the tests themselves though, as it doesn't happen every time? This is something i will investigate but for now, re-running tests causes it to work. There might be a race condition somewhere if it's not anvil.